Load all available glyphs from a font
You can only tell raylib to load a certain amount of codepoints from a font and telling it to try to
load 65,535 codepoints (maximum amount a ttf/otf file can hold) majorly impacts loading time.

Luckily, `std_truetype.h` reports a total amount of glyphs present in a font, so I was able to use
that to load all glyphs from a font.